backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
python-dotenv==1.0.1
neo4j==5.25.0
pypdf==6.6.2
google-genai==0.7.0